Feat: [#44] 작성한 게시글 유효성검사 추가

+++ src/api/board/admin-board.service.ts
@@ -1,6 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { UserEntity } from '../user/entities/user.entity';
 import { BoardResponseDto } from './dtos/boardResponse.dto';
 import { CreateBoardDto } from './dtos/createBoard.dto';
 import { UpdateBoardDto } from './dtos/updateBoard.dto';
@@ -11,12 +16,26 @@ export class AdminBoardService {
   constructor(
     @InjectRepository(AdminBoard)
     private adminBoardRepository: Repository<AdminBoard>,
+    @InjectRepository(UserEntity)
+    private userEntityRepository: Repository<UserEntity>,
   ) {}
 
-  async createBoard(createBoardDto: CreateBoardDto): Promise<number> {
+  async createBoard(
+    createBoardDto: CreateBoardDto,
+    loginUser: number,
+  ): Promise<number> {
+    const user: UserEntity = await this.userEntityRepository.findOne({
+      where: { id: loginUser },
+    });
+
+    if (!user) {
+      throw new NotFoundException('존재하지 않은 회원입니다');
+    }
+
     const board: AdminBoard = new AdminBoard(
       createBoardDto.title,
       createBoardDto.discription,
+      user,
     );
     return (await this.adminBoardRepository.save(board)).id;
   }
@@ -36,19 +55,39 @@ export class AdminBoardService {
   async editBoard(
     id: number,
     updateBoardDto: UpdateBoardDto,
+    loginUser: number,
   ): Promise<BoardResponseDto> {
-    const board: AdminBoard = await this.findById(id);
+    const board: AdminBoard = await this.adminBoardRepository.findOne({
+      where: { id },
+      relations: ['author'],
+    });
+
+    if (board.author.id !== loginUser) {
+      throw new UnauthorizedException(
+        '게시물 등록된 회원의 정보와 로그인정보가 불일치합니다.',
+      );
+    }
+
     const updateBoard = await this.adminBoardRepository.save(
       board.update(updateBoardDto),
     );
     return BoardResponseDto.of(updateBoard);
   }
 
-  async removeBoard(id: number) {
-    const board: AdminBoard = await this.findById(id);
+  async removeBoard(id: number, loginUser: number) {
+    const board: AdminBoard = await this.adminBoardRepository.findOne({
+      where: { id },
+      relations: ['author'],
+    });
+    if (board.author.id !== loginUser) {
+      throw new UnauthorizedException(
+        '게시물 등록된 회원의 정보와 로그인정보가 불일치합니다.',
+      );
+    }
+
     board.delete();
     await this.adminBoardRepository.save(board);
-    return await {
+    return {
       message: '성공적으로 삭제되었습니다.',
     };
   }
